Refactorings of API models namings API responses are now ResourceModels API controller action binding models are now BindingModels
