+++ Quest.API/BindingModels/Quests/CreateSoloInfiniteQuestBM.cs
@@ -1,0 +1,14 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.Quests
+{
+    public class CreateSoloInfiniteQuestBM
+    {
+        [Required]
+        public string Name { get; set; }
+        [Required]
+        public string Description { get; set; }
+        [Required]
+        public string ImageUrl { get; set; }
+    }
+}
