+++ Quest.API/BindingModels/Sessions/LoginBM.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.Sessions
+{
+    public class LoginBM
+    {
+        [Required]
+        public string Username { get; set; }
+        [Required]
+        public string Password { get; set; }
+    }
+}
