+++ Quest.API/BindingModels/SoloPlayers/CreatePlayerBM.cs
@@ -1,0 +1,10 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.SoloPlayers
+{
+    public class CreatePlayerBM
+    {
+        [Required]
+        public int QuestId { get; set; }
+    }
+}
