+++ Quest.API/BindingModels/TaskAttempts/AttemptFeedbackBM.cs
@@ -1,0 +1,8 @@
+namespace Quest.API.BindingModels.TaskAttempts
+{
+    public class AttemptFeedbackBM
+    {
+        public bool IsCorrect { get; set; }
+        public string Comment { get; set; }
+    }
+}
