+++ Quest.API/BindingModels/TaskAttempts/TaskAttemptRM.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.TaskAttempts
+{
+    public class TaskAttemptRM
+    { 
+        [Required] public string AttemptText { get; set; }
+    }
+}
