+++ Quest.API/BindingModels/Teams/AddUserToTeamBM.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.Teams
+{
+    public class AddUserToTeamBM
+    {
+        [Required] public string RequestSecret { get; set; }
+        [Required] public string UserId { get; set; }
+    }
+}
