+++ Quest.API/BindingModels/Teams/AssignModeratorToTeamBM.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.Teams
+{
+    public class AssignModeratorToTeamBM
+    {
+        [Required] public string ModeratorId { get; set; }
+    }
+}
