+++ Quest.API/BindingModels/Teams/CreateTeamBM.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.Teams
+{
+    public class CreateTeamBM
+    {
+        [Required]
+        public int QuestId { get; set; }
+        [Required]
+        [MaxLength(26)]
+        [MinLength(3)]
+        public string Name { get; set; }
+    }
+}
