+++ Quest.API/BindingModels/Users/CreateUserBM.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Quest.API.BindingModels.Users
+{
+    public class CreateUserBM
+    {
+        [Required] public string Username { get; set; }
+        [Required] public string Password { get; set; }
+    }
+}
