+++ Quest.API/Controllers/InvitesController.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Quest.API.Helpers.Errors;
-using Quest.API.ViewModels.Teams;
+using Quest.API.ResourceModels.Participants;
 using Quest.Application.Participants.Commands;
 using Quest.Application.Teams.Commands;
 using Quest.Application.Teams.Queries;
@@ -57,7 +57,7 @@ namespace Quest.API.Controllers
             if (!result.Result)
                 return ApiError.ProblemDetails(HttpStatusCode.Forbidden, result.Message);
             
-            return Ok(new ParticipantDTO(team));
+            return Ok(new ParticipantRM(team));
         }
     }
 }
