+++ Quest.API/Controllers/TeamsController.cs
@@ -10,11 +10,13 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
+using Quest.API.BindingModels.Teams;
 using Quest.API.Helpers;
 using Quest.API.Helpers.Errors;
+using Quest.API.ResourceModels.Participants;
+using Quest.API.ResourceModels.Teams;
+using Quest.API.ResourceModels.Users;
 using Quest.API.Services;
-using Quest.API.ViewModels.Teams;
-using Quest.API.ViewModels.Users;
 using Quest.Application.Participants.Commands;
 using Quest.Application.Services;
 using Quest.Application.Teams.Commands;
@@ -43,20 +45,20 @@ namespace Quest.API.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> GetTeamById(int id)
         {
-            var team = await _mediator.Send(new GetTeamInfoQuery(id));
+            var team = await _mediator.Send(new GetParticipantInfoQuery(id));
 
             if (team == null)
             {
                 return NotFound();
             }
 
-            return Json(new TeamWithCaptainAndMembersDTO(team));
+            return Json(new TeamWithCaptainAndMembersRM(team as Team));
         }
 
 
         [Authorize]
         [HttpPost]
-        public async Task<IActionResult> Post(CreateTeamVM model)
+        public async Task<IActionResult> Post(CreateTeamBM model)
         {
             if (!ModelState.IsValid)
             {
@@ -91,7 +93,7 @@ namespace Quest.API.Controllers
 
         [Authorize]
         [HttpPost("{teamId}/members")]
-        public async Task<IActionResult> AddUserToTeam(int teamId, [FromBody] AddUserToTeamVM model)
+        public async Task<IActionResult> AddUserToTeam(int teamId, [FromBody] AddUserToTeamBM model)
         {
             if (!ModelState.IsValid)
             {
@@ -107,9 +109,9 @@ namespace Quest.API.Controllers
             if (!response.Result)
                 return ApiError.ProblemDetails(HttpStatusCode.Forbidden, response.Message);
 
-            var team = await _mediator.Send(new GetTeamInfoQuery(teamId));
+            var team = await _mediator.Send(new GetParticipantInfoQuery(teamId));
             
-            return Ok(new ParticipantDTO(team));
+            return Ok(new ParticipantRM(team));
         }
 
 
@@ -149,7 +151,7 @@ namespace Quest.API.Controllers
         [HttpPost("{teamId}/moderator/")]
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> AssignModeratorToTeam(int teamId,
-            [FromBody]AssignModeratorToTeamVM model)
+            [FromBody]AssignModeratorToTeamBM model)
         {
             var userId = _userManager.GetUserId(User);
             
@@ -177,7 +179,7 @@ namespace Quest.API.Controllers
             if (response == null)
                 return NotFound();
 
-            return Ok(new ParticipantWithQuestAndModeratorDTO(response));
+            return Ok(new ParticipantWithQuestAndModeratorRM(response));
         }
         
         [Authorize]
@@ -186,12 +188,12 @@ namespace Quest.API.Controllers
         public async Task<IActionResult> GetTeamModerator(int id)
         {
             var response = await _mediator.Send(
-                new GetTeamInfoQuery(id));
+                new GetParticipantInfoQuery(id));
 
             if (response?.Moderator == null)
                 return NotFound();
 
-            return Ok(new ModeratorDTO(response.Moderator));
+            return Ok(new ModeratorRm(response.Moderator));
         }
         
          
@@ -203,15 +205,16 @@ namespace Quest.API.Controllers
             var userId = _userManager.GetUserId(User);
 
             var response = await _mediator.Send(
-                new GetTeamInfoQuery(id));
+                new GetParticipantInfoQuery(id));
 
             if (response == null)
                 return NotFound();
-
-            if (response.Members.All(x => x.UserId != userId))
+            var team = response as Team;
+            
+            if (team.Members.All(x => x.UserId != userId))
                 return ApiError.ProblemDetails(HttpStatusCode.Forbidden, "You are not a member of this team");
             
-            return Ok(response.InviteTokenSecret);
+            return Ok(team.InviteTokenSecret);
         }
     }
 }
