+++ Quest.API/Controllers/UsersController.cs
@@ -16,10 +16,11 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Primitives;
 using Microsoft.IdentityModel.Tokens;
 using NpgsqlTypes;
+using Quest.API.BindingModels.Users;
 using Quest.API.Helpers;
 using Quest.API.Helpers.Errors;
+using Quest.API.ResourceModels.Users;
 using Quest.API.Services;
-using Quest.API.ViewModels.Users;
 using Quest.Application.Users.Commands;
 using Quest.Application.Users.Queries;
 using Quest.DAL.Data;
@@ -55,12 +56,12 @@ namespace Quest.API.Controllers
                 return NotFound("User with that username not found.");
             }
 
-            return Json(new UserDTO(user));
+            return Json(new UserRM(user));
         }
 
         [HttpPost]
         [AllowAnonymous]
-        public async Task<IActionResult> CreateNewUser(CreateUserVM model)
+        public async Task<IActionResult> CreateNewUser(CreateUserBM model)
         {
             var userId = _userManager.GetUserId(User);
             if (userId != null)
