+++ Quest.API/ResourceModels/Participants/ParticipantRM.cs
@@ -1,0 +1,15 @@
+﻿using Quest.Domain.Models;
+
+namespace Quest.API.ResourceModels.Participants
+{
+    public class ParticipantRM
+    {
+        public ParticipantRM(Participant row)
+        {
+            Id = row.Id;
+            Name = row.Name;
+        }
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
