+++ Quest.API/ResourceModels/Users/ModeratorRm.cs
@@ -1,0 +1,14 @@
+using Quest.Domain.Models;
+
+namespace Quest.API.ResourceModels.Users
+{
+    public class ModeratorRm : UserRM
+    {
+        public ModeratorRm(ApplicationUser row) : base(row)
+        {
+            TelegramId = row.TelegramId;
+        }
+        
+        public int TelegramId { get; set; }
+    }
+}
