--- Quest.API/ViewModels/Quests/Create/CreateSoloInfiniteQuestDTO.cs
@@ -1,15 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace Quest.API.ViewModels.Quests
-{
-    public class CreateSoloInfiniteQuestDTO
-    {
-        [Required]
-        public string Name { get; set; }
-        [Required]
-        public string Description { get; set; }
-        [Required]
-        public string ImageUrl { get; set; }
-    }
-}
