--- Quest.API/ViewModels/TaskAttempts/AttemptFeedbackVM.cs
@@ -1,8 +1,0 @@
-namespace Quest.API.ViewModels.TaskAttempts
-{
-    public class AttemptFeedbackVM
-    {
-        public bool IsCorrect { get; set; }
-        public string Comment { get; set; }
-    }
-}
