--- Quest.API/ViewModels/Teams/AssignModeratorToTeamVM.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Quest.API.ViewModels.Teams
-{
-    public class AssignModeratorToTeamVM
-    {
-        [Required] public string ModeratorId { get; set; }
-    }
-}
